test(invitation): moved unit tests to a new file and fixed integration
